python will generate the jwks
