+++ src/core/security.tf
@@ -212,7 +212,7 @@ resource "null_resource" "upload_jwks" {
                 --account-key ${module.checkout_cdn.storage_primary_access_key} \
                 --file "${path.module}/.terraform/tmp/oldJwks.json" \
                 --name '.well-known/jwks.json'
-              python "${path.module}/utils/py/jwkFromPem.py" "${module.jwt.certificate_data_pem}" > "${path.module}/.terraform/tmp/jwks.json"
+              python "${path.module}/utils/py/jwksFromPems.py" "${path.module}/.terraform/tmp/oldJwks.json" "${module.jwt.certificate_data_pem}" > "${path.module}/.terraform/tmp/jwks.json"
               az storage blob upload \
                 --container-name '$web' \
                 --account-name ${replace(replace(module.checkout_cdn.name, "-cdn-endpoint", "-sa"), "-", "")} \
